requirements.txt
PyYAML==	6.0.3
Pygments==	2.19.2
blosc2==	3.12.2
certifi==	2026.1.4
charset-normalizer	==3.4.4
colorama==	0.4.6
delta-spark==	2.4.0
exceptiongroup==	1.3.1
idna==	3.11
importlib_metadata==	8.7.1
iniconfig==	2.3.0
msgpack	==1.1.2
ndindex	==1.10.1
numexpr==	2.14.1
numpy==	2.2.6
packaging==	25.0
pip	==23.2.1
platformdirs	==4.5.1
pluggy==	1.6.0
py-cpuinfo	==9.0.0
py4j	==0.10.9.7
pyspark	==3.4.4
pytest	==9.0.2
requests==	2.32.5
setuptools==	68.2.0
tomli	==2.3.0
typing_extensions	==4.15.0
urllib3==	2.6.2
wheel==	0.41.2
zipp	==3.23.0